Exercise 4.21

+++ part4/blog_list/controllers/blogs.js
@@ -15,6 +15,7 @@ blogsRouter.post("/", async (request, response) => {
 	const body = request.body;
 
 	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	console.log(decodedToken);
 	if (!decodedToken.id) {
 		return response.status(401).json({ error: "token invalid" });
 	}
@@ -37,6 +38,13 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	const blog = await Blog.findById(request.params.id);
+
+	if (decodedToken.id !== blog.user.toString()) {
+		return response.status(401).json({ error: "token invalid" });
+	}
+
 	await Blog.findByIdAndDelete(request.params.id);
 	response.status(204).end();
 });
